refactor: reorganize templates
