+++ lib/src/templates/dart_package/dart_package_bundle.dart
