+++ lib/src/templates/dart_package/dart_package_template.dart
@@ -1,0 +1,30 @@
+import 'package:mason/mason.dart';
+import 'package:universal_io/io.dart';
+import 'package:very_good_cli/src/templates/templates.dart';
+
+/// {@template dart_pkg_template}
+/// A Dart package template.
+/// {@endtemplate}
+class DartPkgTemplate extends Template {
+  /// {@macro dart_pkg_template}
+  DartPkgTemplate()
+      : super(
+          name: 'dart_pkg',
+          bundle: dartPackageBundle,
+          help: 'Generate a reusable Dart package.',
+        );
+
+  @override
+  Future<void> onGenerateComplete(Logger logger, Directory outputDir) async {
+    await installDartPackages(logger, outputDir);
+    await applyDartFixes(logger, outputDir);
+    _logSummary(logger);
+  }
+
+  void _logSummary(Logger logger) {
+    logger
+      ..info('\n')
+      ..alert('Created a Very Good Dart Package! 🦄')
+      ..info('\n');
+  }
+}
