+++ lib/src/templates/flutter_package/flutter_package.dart
@@ -1,0 +1,2 @@
+export 'flutter_package_bundle.dart';
+export 'flutter_package_template.dart';
