+++ lib/src/templates/flutter_package/flutter_package_bundle.dart
