+++ lib/src/templates/flutter_plugin/flutter_plugin.dart
@@ -1,0 +1,2 @@
+export 'flutter_plugin_bundle.dart';
+export 'flutter_plugin_template.dart';
