+++ lib/src/templates/flutter_plugin/flutter_plugin_bundle.dart
