+++ lib/src/templates/very_good_core/very_good_core.dart
@@ -1,0 +1,2 @@
+export 'very_good_core_bundle.dart';
+export 'very_good_core_template.dart';
