+++ lib/src/templates/very_good_core/very_good_core_bundle.dart
