+++ lib/src/templates/very_good_core/very_good_core_template.dart
@@ -1,0 +1,44 @@
+import 'package:io/ansi.dart';
+import 'package:mason/mason.dart';
+import 'package:universal_io/io.dart';
+import 'package:very_good_cli/src/templates/templates.dart';
+
+/// {@template core_template}
+/// A core Flutter app template.
+/// {@endtemplate}
+class CoreTemplate extends Template {
+  /// {@macro core_template}
+  CoreTemplate()
+      : super(
+          name: 'core',
+          bundle: veryGoodCoreBundle,
+          help: 'Generate a Very Good Flutter application.',
+        );
+
+  @override
+  Future<void> onGenerateComplete(Logger logger, Directory outputDir) async {
+    await installFlutterPackages(logger, outputDir);
+    await applyDartFixes(logger, outputDir);
+    _logSummary(logger);
+  }
+
+  void _logSummary(Logger logger) {
+    logger
+      ..info('\n')
+      ..alert('Created a Very Good App! 🦄')
+      ..info('\n')
+      ..info(
+        lightGray.wrap(
+          '''
++----------------------------------------------------+
+| Looking for more features?                         |
+| We have an enterprise-grade solution for companies |
+| called Very Good Start.                            |
+|                                                    |
+| For more info visit:                               |
+| https://verygood.ventures/solution/very-good-start |
++----------------------------------------------------+''',
+        ),
+      );
+  }
+}
